requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
requests
kafka-python
